only delete comments made by the bot

+++ lib/HackTest.rb
@@ -76,9 +76,12 @@ class Hacktest
         pr_payload[:pull_request][:number])
 
     comments.each do |comment|
-      github_client.delete_comment(
-          pr_payload[:repository][:full_name],
-          comment.to_h[:id])
+      comment_hash = comment.to_h
+      if comment_hash[:user][:login] == 'test-app-hackweek-10-20[bot]'
+        github_client.delete_comment(
+            pr_payload[:repository][:full_name],
+            comment_hash[:id])
+      end
     end
   end
